Show plot of the price of USD against VES and require reactable package
I want to show a prettier plot but I still need to figure out how to use
ggplot2 to pass values to an expression. I want to avoid creating a new
column in the main dataset. Also I realized that I required the
reactable package because otherwise I would have to overuse the ::
operator.